mew room support
refs #7

this commit adds in a static response for the "mew" room, as well as
adds in a note about how the web response isn't taken if the peerchat
server returns a special key for GETCHANKEY (which the open one does).
i haven't yet figured out the correct "empty" response for peerchat,
so peerchat will always have priority meaning this endpoint is
basically a no-op.

NOTE: I can generate rooms now, and have figured out most of the fields,
but there are still some i'm missing, and some i have a general idea of
what a byte means, but don't know the range of valid values. to not
overload the maintainers, i'm not going to commit my projects have
generated structures.

in the following weeks I hope to fully sketch out all the fields, all
their values, as well as figure out the "empty" response for peerchat,
and bring those into this project that way they can be used to generate
rooms of any type, and plugged into whatever administration tool mm201
would like to build :)

if there are any rooms i can generate in the meantime, or folks would
like to work on these with me feel free to reach out. :)