Added a folder with all the files I've worked on
I spent too long working on this on my own 😓 It's far from complete, self-evidently; but if we're talking about how close it approaches the drafts I showed you before, it's about 90% done. All the code is heavily commented, both for clarity and to note down what ideas I had in mind for how the website could function.

Here's some questions: If this was the face of the pkmnclassic.net website, would you enjoy visiting it, or do you think it's too gaudy or tacky? Is there anything you wanna add, remove, or rearrange? Since it's currently just regular ol' HTML, how will it be converted for ASP.NET? Can this be effectively revised for accessibility? Do you hate how much I relied on tables for layout purposes? Whaddaya think???